+++ packages/chain-events/cosmos-chain-testing/README.md
@@ -1,12 +1,18 @@
 # Cosmos Chain Testing Tools
 
-This creates a sandbox Cosmos SDK devnet node, which is a companion to the `http://localhost:8080/csdk` sandbox community. This is running v0.46.11 of Cosmos SDK. This is our first support level for `v1` of the the `gov` module, so it is useful for testing v1 changes.
+This creates two local Cosmos SDK devnet nodes. `csdk-v1` is running v0.46.11 of Cosmos SDK. `csdk-beta-ci` is running v0.45.0. We can use these to manual test and run the tests for `yarn test-devnet`.
 
 ## Setup 
 
-The easiest way to access the tool is to run `yarn cosmos:build` this will create a docker deployment with two endpoints. The following networking configuration is as follows:
-1. RPC - `localhost:5050/rpc`
-2. LCD (REST API) - `localhost:5050/lcd/`
+The easiest way to access the tool is to run `yarn cosmos:build` this will create a docker deployment with two devnets, with two endpoints each. The networking configuration is as follows:
+
+gov v1:
+1. RPC - `http://localhost:5050/rpc`
+2. LCD (REST API) - `http://localhost:5050/lcd/`
+
+gov v1beta1:
+1. RPC - `http://localhost:5051/rpc`
+2. LCD (REST API) - `http://localhost:5051/lcd/`
 
 ## Testing
 
@@ -14,11 +20,18 @@ Once you have a testnet running in docker, you can:
     - test the UI by importing this mnemonic into keplr:
         - 'ignore medal pitch lesson catch stadium victory jewel first stairs humble excuse
           scrap clutch cup daughter bench length sell goose deliver critic favorite thought'
-    - run the tests in `packages/commonwealth/test/integration/api/external/proposalTx.spec.ts`
+    - run the tests in `packages/commonwealth/test/devnet/proposalTx.spec.ts` with `yarn test-devnet`
 
-## Live node
+## Live nodes
 
+These are sandbox communities for manual testing
 See [wiki](https://github.com/hicommonwealth/commonwealth/wiki/Devnet)
+
+gov v1:
 -  https://cosmos-devnet.herokuapp.com/rpc
 -  https://cosmos-devnet.herokuapp.com/lcd/
+
+gov v1beta1:
+-  https://cosmos-devnet-beta.herokuapp.com/rpc
+-  https://cosmos-devnet-beta.herokuapp.com/lcd/
