+++ packages/chain-events/cosmos-chain-testing/v1/nginx.conf
