+++ packages/chain-events/cosmos-chain-testing/v1/stop.sh
