+++ packages/chain-events/cosmos-chain-testing/v1beta1/heroku.yml
@@ -1,0 +1,3 @@
+build:
+  docker:
+    web: dockerfile
