+++ packages/chain-events/cosmos-chain-testing/v1beta1/nginx.conf
@@ -1,0 +1,19 @@
+events {
+  worker_connections  1024;
+}
+
+http {
+  server {
+    listen 80;
+    server_name localhost;
+
+    location /rpc {
+      proxy_pass http://127.0.0.1:26657/;
+    }
+
+    location /lcd/ {
+      proxy_pass http://127.0.0.1:1317/;
+    }
+
+  }
+}
