+++ packages/chain-events/cosmos-chain-testing/v1beta1/start.sh
@@ -5,11 +5,11 @@ set -x
 SCRIPT_NAME=$(basename "$0")
 SCRIPT_DIR=$(cd "$(dirname "$0")" && pwd)
 
-PORT=5050
-IMAGE_NAME=heroku-csdk
+PORT=5051
+IMAGE_NAME=csdk-beta
 
 if [ "$1" = "--build" ]; then
     cd $SCRIPT_DIR
     docker build -t ${IMAGE_NAME} .
 fi
-docker run -p 5050:5050 -e PORT=${PORT} --name ${IMAGE_NAME} ${IMAGE_NAME}
+docker run -d -p 5051:5051 -e PORT=${PORT} --name ${IMAGE_NAME} ${IMAGE_NAME}
