+++ packages/chain-events/cosmos-chain-testing/v1beta1/stop.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# This script is used to start a local cosmos chain for testing purposes.
+
+set -x
+IMAGE_NAME=csdk-beta
+# docker force stop and remove the container but leave the image
+docker stop ${IMAGE_NAME} && docker rm ${IMAGE_NAME} -f || trues
