+++ packages/chain-events/package.json
@@ -43,9 +43,9 @@
     "dump-db": "pg_dump $(heroku config:get DATABASE_URL -a chain-events) --verbose --no-privileges --no-owner -f latest.dump",
     "dump-db-empty": "pg_dump $(heroku config:get DATABASE_URL -a chain-events) --exclude-table-data=\"public.\\\"ChainEntities\\\"\" --exclude-table-data=\"public.\\\"ChainEventTypes\\\"\" --exclude-table-data=\"public.\\\"ChainEvents\\\"\" --verbose --no-privileges --no-owner -f latest.dump",
     "migrate-events": "NODE_OPTIONS=--max-old-space-size=$(../../scripts/get-max-old-space-size.sh) ts-node --project tsconfig.json scripts/migrateChainEntities.ts run-as-script",
-    "cosmos:build": "chmod u+x cosmos-chain-testing/start.sh && ./cosmos-chain-testing/start.sh --build",
-    "cosmos:start": "chmod u+x cosmos-chain-testing/start.sh && ./cosmos-chain-testing/start.sh",
-    "cosmos:stop": "chmod u+x cosmos-chain-testing/stop.sh && ./cosmos-chain-testing/stop.sh"
+    "cosmos:build": "chmod u+x cosmos-chain-testing/v1/start.sh && ./cosmos-chain-testing/v1/start.sh --build && chmod u+x cosmos-chain-testing/v1beta1/start.sh && ./cosmos-chain-testing/v1beta1/start.sh --build",
+    "cosmos:start": "chmod u+x cosmos-chain-testing/v1/start.sh && ./cosmos-chain-testing/v1/start.sh && chmod u+x cosmos-chain-testing/v1beta1/start.sh && ./cosmos-chain-testing/v1beta1/start.sh",
+    "cosmos:stop": "chmod u+x cosmos-chain-testing/v1/stop.sh && ./cosmos-chain-testing/v1/stop.sh && chmod u+x cosmos-chain-testing/v1beta1/stop.sh && ./cosmos-chain-testing/v1beta1/stop.sh"
   },
   "engines": {
     "node": "18.x"
@@ -118,7 +118,7 @@
     "install-peers-cli": "^2.2.0",
     "jsdom": "^16.2.2",
     "jsdom-global": "^3.0.2",
-    "mocha": "^8.2.1",
+    "mocha": "^10.0.0",
     "nyc": "^15.1.0",
     "prettier": "2.0.2",
     "rimraf": "^3.0.2",
