+++ packages/commonwealth/package.json
@@ -266,7 +266,7 @@
     "throttled-queue": "^1.0.7",
     "ts-node": "^8.6.2",
     "tsc-alias": "^1.7.0",
-    "tsconfig-paths": "^3.8.0",
+    "tsconfig-paths": "^4.2.0",
     "tweetnacl": "^1.0.3",
     "underscore": "^1.13.1",
     "usehooks-ts": "^2.9.1",
@@ -315,7 +315,7 @@
     "nyc": "^15.1.0",
     "playwright-test": "^8.2.0",
     "process": "^0.11.10",
-    "source-map-support": "^0.5.13",
+    "source-map-support": "^0.5.21",
     "storybook": "7.0.9",
     "sync-request": "^6.1.0",
     "ts-mocha": "^6.0.0",
